pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "privibot"
version = "0.1.1"
description = "Privilege system for Telegram bots."
authors = ["Timothée Mazzucotelli <pawamoy@pm.me>"]
license = "ISC License"
readme = "README.md"
repository = "https://github.com/pawamoy/privibot"
homepage = "https://github.com/pawamoy/privibot"
keywords = ["telegram-bot"]
packages = [ { include = "privibot", from = "src" } ]
include = [
    "README.md",
    "pyproject.toml"
]

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = { version = "=12.0.0b1", allow-prereleases = true }
sqlalchemy = "^1.3"

coverage = {version = "^5.2.1", optional = true}
invoke = {version = "^1.4.1", optional = true}
mypy = {version = "^0.782", optional = true}
pytest = {version = "^6.0.1", optional = true}
pytest-randomly = {version = "^3.4.1", optional = true}
pytest-sugar = {version = "^0.9.4", optional = true}
pytest-xdist = {version = "^2.1.0", optional = true}

[tool.poetry.extras]
tests = ["coverage", "invoke", "mypy", "pytest", "pytest-randomly", "pytest-sugar", "pytest-xdist"]

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^20.8b1"
coverage = "^5.2.1"
failprint = "^0.3.0"
flakehell = "^0.6.0"
flake8-black = "^0.2.1"
flake8-builtins = "^1.5.3"
flake8-tidy-imports = "^4.1.0"
flake8-variables-names = "^0.0.3"
flake8-pytest-style = "^1.3.0"
git-changelog = "^0.4.0"
httpx = "^0.14.3"
invoke = "^1.4.1"
ipython = "^7.2"
# TODO: accept v5 once https://github.com/wemake-services/wemake-python-styleguide/issues/1584 is resolved
isort = {version = "<5", extras = ["pyproject"]}
jinja2-cli = "^0.7.0"
mkdocs = "^1.1.2"
mkdocs-material = "^5.5.12"
mkdocstrings = "^0.13.1"
mypy = "^0.782"
pytest = "^6.0.1"
pytest-randomly = "^3.4.1"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.1.0"
toml = "^0.10.1"
wemake-python-styleguide = "^0.14.1"
# TODO: remove once https://github.com/python-poetry/poetry/issues/2728 is resolved
wrapt = "^1.12.1"

[tool.poetry.scripts]
privibot = "privibot.cli:main"

[tool.black]
line-length = 120
exclude = "tests/fixtures"

[tool.isort]
line_length = 120
not_skip = "__init__.py"
multi_line_output = 3
force_single_line = false
balanced_wrapping = true
default_section = "THIRDPARTY"
known_first_party = "privibot"
include_trailing_comma = true

[tool.flakehell]
format = "colored"
max_line_length = 132
show_source = false
exclude = ["tests/fixtures"]

[tool.flakehell.plugins]
"*" = [
    "+*",
    "-RST*",  # we write docstrings in markdown, not rst
    "-A001",  # redundant with W0622 (builtin override), which is more precise about line number
    "-D105",  # missing docstring in magic method
    "-D212",  # multi-line docstring summary should start at the first line
    "-E203",  # whitespace before ‘:’ (incompatible with Black)
    "-Q000",  # black already deals with quoting
    "-S101",  # use of assert
    "-W503",  # line break before binary operator (incompatible with Black)
    "-C0116",  # redunant with D102 (missing docstring)
    "-C0301",  # line too long
    "-R0903",  # too few public methods
    "-R0913",  # too many methods
    "-R0914",  # too many local variables
    "-W0611",  # redundant with F401 (unused import)
    "-WPS305",  # f-strings
    "-WPS125",  # redundant with W0622 (builtin override), which is more precise about line number
    "-WPS202",  # too many module members
    "-WPS210",  # too many local variables
    "-WPS213",  # too many expressions
    "-WPS220",  # too deep nesting
    "-WPS226",  # string over-use: can't disable it per file?
    "-WPS326",  # implicit string concatenation
    "-WPS336",  # explicit string concatenation
    "-WPS412",  # __init__ modules with logic
    "-WPS433",  # redundant with C0415 (not top-level import)
]
